% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_all.R
\name{RC}
\alias{RC}
\title{Perform Regression Calibration for all estimators (IVW, simple/weighted median, simple/weighted mode)}
\usage{
RC(dat)
}
\arguments{
\item{dat}{Harmonised exposure (discovery and replication) and outcome data. Output from \code{TwoSampleMR::harmonise_data}.
Note than one extra column is needed (\code{sample}) and than only the following columns are used:
\code{SNP}, \code{beta.exposure}, \code{beta.outcome}, \code{se.exposure}, \code{se.outcome} and \code{mr_keep}}
}
\value{
data.frame with Regression Calibration results
For each method (IVW, simple/weighted median, simple/weighted mode)
\describe{
\item{nsnps}{Number of instruments used}
\item{b}{Causal effect estimate}
\item{se}{Causal effect standard error}
\item{pval}{Causal effect p-value}
}
}
\description{
Perform Regression Calibration for all estimators (IVW, simple/weighted median, simple/weighted mode)
}
