% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_all.R
\name{RC_simple_mode}
\alias{RC_simple_mode}
\title{Simple Mode Regression Calibration}
\usage{
RC_simple_mode(b_disc, b_rep, b_out, se_disc, se_rep, se_out)
}
\arguments{
\item{b_disc}{SNP-exposure association estimate in discovery sample}

\item{b_rep}{SNP-exposure association estimate in replication sample}

\item{b_out}{SNP-outcome association estimate}

\item{se_disc}{SNP-exposure association standard error in discovery sample}

\item{se_rep}{SNP-exposure association standard error in replication sample}

\item{se_out}{SNP-outome association standard error}
}
\value{
List with the following elements:
\describe{
\item{b}{Causal effect estimate}
\item{se}{Causal effect standard error}
\item{pval}{Causal effect p-value}
\item{nsnps}{Number of instruments used}
}
}
\description{
Simple Mode Regression Calibration
}
